tests/test_pg_connection.py
# test_pg_connection.py
import os
import psycopg2
from dotenv import load_dotenv

load_dotenv()

DB_HOST = os.getenv("DB_HOST", "localhost")
DB_PORT = os.getenv("DB_PORT", "8000")
DB_NAME = os.getenv("DB_DB", "postgres")
DB_USER = os.getenv("DB_USER", "postgres")
DB_PASS = os.getenv("DB_PASSWORD", "password")

try:
    conn = psycopg2.connect(
        host=DB_HOST,
        port=DB_PORT,
        dbname=DB_NAME,
        user=DB_USER,
        password=DB_PASS,
    )
    cur = conn.cursor()
    cur.execute("SELECT version();")
    version = cur.fetchone()
    print(f"✅ Connected to PostgreSQL: {version[0]}")
    
    cur.execute("SELECT table_name FROM information_schema.tables WHERE table_schema='public';")
    tables = cur.fetchall()
    print("📦 Tables in 'public' schema:")
    for table in tables:
        print(f" - {table[0]}")

    cur.close()
    conn.close()

except Exception as e:
    print(f"❌ Connection failed: {e}")
