requirements.txt
fastapi
uvicorn
python-dotenv
boto3
qdrant-client
